fix(Table): corrections finalized
